Move from SvelteKit to Astro for static site generation + styling for mobile
