+++ .gitignore
@@ -1,21 +1,24 @@
-node_modules
+# build output
+dist/
+# generated types
+.astro/
 
-# Output
-.output
-.vercel
-/.svelte-kit
-/build
+# dependencies
+node_modules/
 
-# OS
-.DS_Store
-Thumbs.db
+# logs
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
 
-# Env
+
+# environment variables
 .env
-.env.*
-!.env.example
-!.env.test
+.env.production
 
-# Vite
-vite.config.js.timestamp-*
-vite.config.ts.timestamp-*
+# macOS-specific files
+.DS_Store
+
+# jetbrains setting folder
+.idea/
