--- .prettierignore
@@ -1,4 +1,0 @@
-# Package Managers
-package-lock.json
-pnpm-lock.yaml
-yarn.lock
