+++ astro.config.mjs
@@ -1,0 +1,9 @@
+// @ts-check
+import { defineConfig } from 'astro/config';
+
+import svelte from '@astrojs/svelte';
+
+// https://astro.build/config
+export default defineConfig({
+    integrations: [svelte()],
+});
