+++ src/data/selections.json
@@ -1,0 +1,166 @@
+[
+  {
+    "player": "Alex Jones",
+    "draftPosition": 1,
+    "picks": [
+      {
+        "pickNumber": 1,
+        "team": "Washington Wizards",
+        "selection": "L"
+      },
+      {
+        "pickNumber": 2,
+        "team": "Dallas Mavericks",
+        "selection": "W"
+      },
+      {
+        "pickNumber": 3,
+        "team": "Utah Jazz",
+        "selection": "L"
+      },
+      {
+        "pickNumber": 4,
+        "team": "Cleveland Cavaliers",
+        "selection": "W"
+      },
+      {
+        "pickNumber": 5,
+        "team": "Philadelphia 76ers",
+        "selection": "W"
+      },
+      {
+        "pickNumber": 6,
+        "team": "New Orleans Pelicans",
+        "selection": "W"
+      },
+      {
+        "pickNumber": 7,
+        "team": "Indiana Pacers",
+        "selection": "W"
+      }
+    ]
+  },
+  {
+    "player": "Dylan Cawsey",
+    "draftPosition": 2,
+    "picks": [
+      {
+        "pickNumber": 1,
+        "team": "Portland Trail Blazers",
+        "selection": "L"
+      },
+      {
+        "pickNumber": 2,
+        "team": "Toronto Raptors",
+        "selection": "L"
+      },
+      {
+        "pickNumber": 3,
+        "team": "Minnesota Timberwolves",
+        "selection": "W"
+      },
+      {
+        "pickNumber": 4,
+        "team": "LA Clippers",
+        "selection": "L"
+      },
+      {
+        "pickNumber": 5,
+        "team": "Phoenix Suns",
+        "selection": "W"
+      },
+      {
+        "pickNumber": 6,
+        "team": "Orlando Magic",
+        "selection": "W"
+      },
+      {
+        "pickNumber": 7,
+        "team": "Golden State Warriors",
+        "selection": "W"
+      }
+    ]
+  },
+  {
+    "player": "Zachary Dries",
+    "draftPosition": 3,
+    "picks": [
+      {
+        "pickNumber": 1,
+        "team": "Oklahoma City Thunder",
+        "selection": "W"
+      },
+      {
+        "pickNumber": 2,
+        "team": "Detroit Pistons",
+        "selection": "L"
+      },
+      {
+        "pickNumber": 3,
+        "team": "New York Knicks",
+        "selection": "W"
+      },
+      {
+        "pickNumber": 4,
+        "team": "Denver Nuggets",
+        "selection": "W"
+      },
+      {
+        "pickNumber": 5,
+        "team": "Milwaukee Bucks",
+        "selection": "W"
+      },
+      {
+        "pickNumber": 6,
+        "team": "San Antonio Spurs",
+        "selection": "L"
+      },
+      {
+        "pickNumber": 7,
+        "team": "Houston Rockets",
+        "selection": "L"
+      }
+    ]
+  },
+  {
+    "player": "Jack Peterkin",
+    "draftPosition": 4,
+    "picks": [
+      {
+        "pickNumber": 1,
+        "team": "Brooklyn Nets",
+        "selection": "L"
+      },
+      {
+        "pickNumber": 2,
+        "team": "Boston Celtics",
+        "selection": "W"
+      },
+      {
+        "pickNumber": 3,
+        "team": "Charlotte Hornets",
+        "selection": "L"
+      },
+      {
+        "pickNumber": 4,
+        "team": "Chicago Bulls",
+        "selection": "L"
+      },
+      {
+        "pickNumber": 5,
+        "team": "Sacramento Kings",
+        "selection": "W"
+      },
+      {
+        "pickNumber": 6,
+        "team": "Atlanta Hawks",
+        "selection": "L"
+      },
+      {
+        "pickNumber": 7,
+        "team": "Memphis Grizzlies",
+        "selection": "W"
+      }
+    ]
+  }
+]
