--- src/lib/types/team.ts
@@ -1,5 +1,0 @@
-export interface TeamData {
-	teamNickname: string;
-	win: number;
-	loss: number;
-}
