+++ src/pages/index.astro
@@ -1,0 +1,16 @@
+---
+import ResultsBoard from '../components/ResultsBoard.svelte';
+---
+
+<html lang='en'>
+    <head>
+        <meta charset='utf-8' />
+        <link rel='icon' type='image/svg+xml' href='/favicon.svg' />
+        <meta name='viewport' content='width=device-width' />
+        <meta name='generator' content={Astro.generator} />
+        <title>2024 NBA Skins Draft</title>
+    </head>
+    <body>
+        <ResultsBoard />
+    </body>
+</html>
