+++ src/types/state.ts
@@ -1,5 +1,5 @@
 import type { DraftSelectionsWithTeamData } from './draft';
 
 export interface HomePageData {
-	draftSelectionsWithTeamData: DraftSelectionsWithTeamData[];
+    draftSelectionsWithTeamData: DraftSelectionsWithTeamData[];
 }
