+++ src/types/team.ts
@@ -1,0 +1,5 @@
+export interface TeamData {
+    teamNickname: string;
+    win: number;
+    loss: number;
+}
